更新代码注释，简单做了代码的修改和优化
