--- src/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-# @Time    : 2018/5/26 下午5:16
-# @Author  : zhanzecheng
-# @File    : __init__.py.py
-# @Software: PyCharm
-"""
